+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Add css and javascript content types
+
 ### 0.27.2 - 2020-08-31
 
  - Fix background_task module